# cmd/server/docs/swagger.yaml
definitions:
  github_com_qhmd_gitforgits_internal_domain_auth.Auth:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  users.EmailAlreadyUsed:
    properties:
      error:
        example: this is email already used
        type: string
      message:
        example: Email already exist, choose another Email
        type: string
      success:
        example: false
        type: boolean
    type: object
  users.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
      success:
        example: false
        type: boolean
    type: object
  users.InvalidId:
    properties:
      error:
        example: Invalid id
        type: string
      message:
        example: Your id is invalid
        type: string
      success:
        example: false
        type: boolean
    type: object
  users.SuccessDeleteUser:
    properties:
      message:
        example: success delete user
        type: string
      success:
        example: true
        type: boolean
    type: object
  users.SuccessGetList:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_qhmd_gitforgits_internal_domain_auth.Auth'
        type: array
      message:
        example: successfully get list user
        type: string
      success:
        example: true
        type: boolean
    type: object
  users.SuccessGetUser:
    properties:
      data:
        $ref: '#/definitions/github_com_qhmd_gitforgits_internal_domain_auth.Auth'
      message:
        example: successfully get the user
        type: string
      success:
        example: true
        type: boolean
    type: object
  users.UpdateRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 50
        minLength: 4
        type: string
      password:
        type: string
      role:
        enum:
        - user
        - admin
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
  users.UserNotFoundResponse:
    properties:
      errors:
        example: user with id {id} does not exist
        type: string
      message:
        example: user not found
        type: string
      success:
        example: false
        type: boolean
    type: object
info:
  contact: {}
  description: |-
    API documentation for project GitForGits
    Login as admin:
    email: admingitforgits12@gmail.com
    password: @GitForGitsAdmin21
  termsOfService: http://swagger.io/terms/
  title: GitForGits API
  version: "1.0"
paths:
  /admin/users/:
    get:
      consumes:
      - application/json
      description: Get list user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.SuccessGetList'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/users.UserNotFoundResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get List User
      tags:
      - Users
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.SuccessDeleteUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/users.InvalidId'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/users.UserNotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/users.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve a single user by its ID
      parameters:
      - description: Users ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.SuccessGetUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/users.InvalidId'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/users.UserNotFoundResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Account
        in: body
        name: Auth
        required: true
        schema:
          $ref: '#/definitions/users.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.SuccessGetList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/users.InvalidId'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/users.UserNotFoundResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/users.EmailAlreadyUsed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/users.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"